% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating-packages.R
\docType{package}
\name{branchestimate-package}
\alias{branchestimate}
\alias{branchestimate-package}
\title{branchestimate: Malthusian parameter and R effective estimator for US regional COVID-19 data}
\description{
Estimates the Malthusian parameters and R effective in regional US data for those regions experiencing epidemic COVID-19 growth.
}
\author{
\strong{Maintainer}: Jérôme Levesque \email{jerome.levesque@tpsgc-pwgsc.gc.ca}

Authors:
\itemize{
  \item David Maybury \email{david.maybury@tpsgc-pwgsc.gc.ca}
}

}
\keyword{internal}
