% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch-estimating-functions.R
\name{make_county_choro_data}
\alias{make_county_choro_data}
\title{Creates a tibble for choroplethR function county_choropleth.}
\usage{
make_county_choro_data(mcmc_object)
}
\arguments{
\item{mcmc_object}{The list provided by run_jags_estimator() for county regions.}
}
\value{
A tibble for county_choropleth
}
\description{
This function creates the tibble that county_choropleth() requires in the correct format. The value
is the R effective for the region.
}
