% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch-estimating-functions.R
\name{make_state_choro_data}
\alias{make_state_choro_data}
\title{Creates a tibble for choroplethR function state_choropleth.}
\usage{
make_state_choro_data(mcmc_object)
}
\arguments{
\item{mcmc_object}{The list provided by run_jags_estimator() for state regions.}
}
\value{
A tibble for state_choropleth
}
\description{
This function creates the tibble that state_choropleth() requires in the correct format. The value
is the R effective for the region
}
