% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch-estimating-functions.R
\name{run_jag_estimator}
\alias{run_jag_estimator}
\title{A JAGS wrapper for Malthusian parameter estimation based on a regional random effect.}
\usage{
run_jag_estimator(master_data_2)
}
\arguments{
\item{master_data_2}{The result of US_covid_data_creator(): county or state regions.}
}
\value{
A list containing a BUGS mcmc object and the original data.
}
\description{
This function creates the MCMC posterior output of JAGS running on the US regional COVID-19 data.
The output is based on a linear mixed model of the log transformed count data providing both the fixed effects
and the random effects.
}
