% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branch-estimating-functions.R
\name{US_covid_data_creator}
\alias{US_covid_data_creator}
\title{Extractor of COVID-19 data from the NYT.}
\usage{
US_covid_data_creator(
  date_start = "2020-05-01",
  date_end = "2020-08-20",
  county = TRUE
)
}
\arguments{
\item{date_start}{The start date of the time window. Default 2020-05-01}

\item{date_end}{The end date of the time window. Default 2020-08-20}

\item{county}{A logical TRUE = county level data; FALSE = state level data}
}
\value{
A tibble containing the regional data including fips number for county level data.
}
\description{
This function extracts county/state level COVID-19 cumulative count data from the NYT: https://raw.githubusercontent.com/nytimes/covid-19-data/master/us-counties.csv
The function filters only those regions that exhibit exponential growth. The filtering occurs through a sign check
on a quadratic regression over the selected time window.
}
